requirements.txt
fastapi
uvicorn
pydantic
python-multipart
resend
python-dotenv
google-generativeai
